requirements.txt
streamlit
pandas
pillow
requests
python-jose
streamlit-cookies-manager
openpyxl
pyarrow
fastapi
uvicorn
python-dotenv
pytesseract
openai
authlib
sqlalchemy
psycopg2-binary
numpy
python-multipart
starlette